pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-revenue-analyzer"
version = "0.1.0"
description = "LLM cost and revenue analytics API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "sqlalchemy>=2.0.32,<3.0",
  "alembic>=1.13,<2.0",
  "psycopg[binary]>=3.2,<4.0",
  "PyYAML>=6.0.1,<7.0",
  "prometheus-client>=0.20,<1.0",
  "python-dateutil>=2.9,<3.0",
  "httpx>=0.27,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "ruff>=0.6.9,<1.0",
  "mypy>=1.11,<2.0",
  "pre-commit>=3.8,<4.0",
  "types-PyYAML>=6.0.12.20240917",
]

[project.scripts]
llm-revenue-api = "llm_revenue_analyzer.main:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["src/app", "scripts", "alembic"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
packages = ["llm_revenue_analyzer"]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["llm_revenue_analyzer.scripts.*"]
ignore_errors = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
